% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_leaf_traits.r
\name{calculate_leaf_traits}
\alias{calculate_leaf_traits}
\title{Calculates leaf traits for a directory of leaf scans
either made with a white or black background.}
\usage{
calculate_leaf_traits(path = "~", dpi = 300, background = "white",
  min_size = 10000, reference = c(1000, 1000), out_path = NULL,
  plot = TRUE)
}
\arguments{
\item{path}{path of image files to process (should only contain images),
default is your home directory}

\item{dpi}{dots per inch resolution of the scan (default = 300)}

\item{background}{background matte used ("white" or "black")}

\item{min_size}{minimum size of a patch (in pixels) to be considered valid
take into consideration that this is resolution dependent (default = 10000)}

\item{reference}{is there a panel to the left which
should be excluded, provide a number of pixels to do so (default = c(100,100))}

\item{out_path}{output directory where to save the data if not returned
to an R variable}

\item{plot}{plot the processed images for reference (default = FALSE)}
}
\value{
A nested list (data structure) listing the extracting traits together
with the filename of the processed file.
}
\description{
Calculates leaf traits for a directory of leaf scans
either made with a white or black background.
}
\examples{
\dontrun{

leaf_statistics = calculate_leaf_traits(path = "/mydata")

}
}
\keyword{Gcc,}
\keyword{Rcc}
\keyword{area,}
\keyword{leaf}
\keyword{specific}
\keyword{traits,}
